requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cloudinary==1.44.0
dj-database-url==2.3.0
Django==4.2.20
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dotenv==0.9.9
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
gunicorn==21.2.0
whitenoise==6.6.0
